server/requirements.txt
Flask
Flask-Caching
requests
psycopg2-binary
Flask-Cors
python-dotenv